refactor: split the data and the view layer fo the data table
