+++ frontend/src/stores/index.ts
@@ -1,0 +1,5 @@
+import useDBInstanceStore from "./dbInstance";
+import useSearchConfigStore from "./searchConfig";
+import useDataTableStore from "./dataTable";
+
+export { useDBInstanceStore, useSearchConfigStore, useDataTableStore };
