+++ frontend/src/views/AppView.vue
@@ -38,10 +38,10 @@
   <div class="mx-5 mt-4 border-b pb-4">
     <!-- selected dashboard -->
     <cost-table-checked
-      :data-row="state.checkedDataRow"
+      :data-row="dataTableStore.checkedDataRow"
       :is-loading="state.isLoading"
       :is-expended="state.isCheckedTableExpended"
-      :checked-row-keys="state.checkRowKeys"
+      :checked-row-keys="dataTableStore.checkedRowKey"
       @update-checked-row-keys="
         (val:string[]) => {
           handleCheckRowKeys(val, false);
@@ -54,7 +54,7 @@
     <div class="mt-2">
       <chart-tab
         v-show="state.isCheckedTableExpended"
-        :data="state.checkedDataRow"
+        :data="dataTableStore.checkedDataRow"
       />
     </div>
   </div>
@@ -62,11 +62,11 @@
   <!-- dashboard -->
   <div class="mx-5 mt-5">
     <cost-table
-      :data-row="state.dataRow"
+      :data-row="dataTableStore.dataRow"
       :is-loading="state.isLoading"
       :show-engine-type="showEngineType"
       :show-lease-length="showLeaseLength"
-      :checked-row-keys="state.checkRowKeys"
+      :checked-row-keys="dataTableStore.checkedRowKey"
       @update-checked-row-keys="
         (val:string[]) => {
           handleCheckRowKeys(val, true);
@@ -78,8 +78,6 @@
 
 <script setup lang="ts">
 import { watch, reactive, ref, computed, onMounted } from "vue";
-
-import { DataRow } from "../components/CostTable";
 import { NButton, useNotification } from "naive-ui";
 
 import {
@@ -88,21 +86,26 @@ import {
   CloudProvider,
   EngineType,
 } from "../types";
-import { useDBInstanceStore } from "../stores/dbInstance";
-import { useSearchConfigStore } from "../stores/searchConfig";
+import {
+  useSearchConfigStore,
+  useDBInstanceStore,
+  useDataTableStore,
+} from "../stores";
 import { useRouter } from "vue-router";
 
 import { RouteParam } from "../router";
-import { isEmptyArray, getRegionCode, getRegionName } from "../util";
+import { isEmptyArray } from "../util";
 
 const dbInstanceStore = useDBInstanceStore();
 dbInstanceStore.loadDBInstanceList();
+
+const dataTableStore = useDataTableStore();
 
 watch(
   () => dbInstanceStore.dbInstanceList.length,
   () => {
     state.availableRegions = dbInstanceStore.getAvailableRegionList();
-    refreshDataTable();
+    dataTableStore.refresh();
   }
 );
 
@@ -110,18 +113,12 @@ const searchConfigStore = useSearchConfigStore();
 
 interface LocalState {
   availableRegions: AvailableRegion[];
-  dataRow: DataRow[];
-  checkRowKeys: string[];
-  checkedDataRow: DataRow[];
   isLoading: boolean;
   isCheckedTableExpended: boolean;
 }
 
 const state = reactive<LocalState>({
   availableRegions: [],
-  dataRow: [],
-  checkRowKeys: [],
-  checkedDataRow: [],
   isLoading: false,
   isCheckedTableExpended: false,
 });
@@ -148,23 +145,15 @@ const handleUpdateMinCPU = (val: any) => {
   searchConfigStore.searchConfig.minCPU = val;
 };
 const handleCheckRowKeys = (rowKeys: string[], needScroll: boolean) => {
-  const rowKeySet = new Set<string>(rowKeys);
-  state.checkedDataRow = state.dataRow.filter((row: DataRow) =>
-    rowKeySet.has(row.key)
-  );
-  state.checkedDataRow.forEach((row) => {
-    row.childCnt = 1;
-  });
+  dataTableStore.refreshChecked(rowKeys);
 
   if (state.isCheckedTableExpended && needScroll) {
-    if (state.checkRowKeys.length > rowKeys.length) {
+    if (dataTableStore.checkedRowKey.length > rowKeys.length) {
       window.scrollBy(0, -47);
     } else {
       window.scrollBy(0, 47);
     }
   }
-
-  state.checkRowKeys = rowKeys;
 };
 
 const handleToggleIsExpanded = () => {
@@ -228,10 +217,10 @@ watch(
   () => {
     // We set the dataRow to empty here for a better animation.
     // Otherwise the loading circle would appear right in the middle of the data table, which may be elusive.
-    state.dataRow = [];
+    dataTableStore.clearDataRow();
     state.isLoading = true;
     setTimeout(() => {
-      refreshDataTable();
+      dataTableStore.refresh();
       state.isLoading = false;
     }, 100);
   },
@@ -247,141 +236,15 @@ const showLeaseLength = computed(() => {
   const chargeTypeSet = new Set(config.value.chargeType);
   return chargeTypeSet.size > 1 || chargeTypeSet.has("Reserved");
 });
-const refreshDataTable = () => {
-  let rowCnt = 0;
-  const config = searchConfigStore.searchConfig;
-  const dbInstanceList = dbInstanceStore.dbInstanceList;
-
-  if (!config.region && !config.engineType && !config.chargeType) {
-    return;
-  }
-
-  const regionCodeList: string[] = [];
-  config.region?.forEach((regionName) => {
-    regionCodeList.push(...getRegionCode(regionName));
-  });
-  const cloudProviderSet = new Set(config.cloudProvider);
-  const selectedRegionCodeSet = new Set(regionCodeList);
-  const engineSet = new Set(config.engineType);
-  const chargeTypeSet = new Set(config.chargeType);
-
-  dbInstanceList.forEach((dbInstance) => {
-    if (
-      (config.minRAM && Number(dbInstance.memory) < config.minRAM) ||
-      (config.minCPU && Number(dbInstance.cpu) < config.minCPU)
-    ) {
-      return;
-    }
-
-    if (!cloudProviderSet.has(dbInstance.cloudProvider)) {
-      return;
-    }
-
-    const selectedRegionList = dbInstance.regionList.filter((region) => {
-      if (selectedRegionCodeSet.has(region.code)) {
-        return true;
-      }
-      return false;
-    });
-    // no region found
-    if (selectedRegionList.length === 0) {
-      return;
-    }
-
-    const dataRowList: DataRow[] = [];
-    const dataRowMap: Map<string, DataRow[]> = new Map();
-
-    // filter terms provided in each region
-    selectedRegionList.forEach((region) => {
-      const selectedTermList = region.termList.filter((term) => {
-        if (
-          chargeTypeSet.has(term.type) &&
-          engineSet.has(term.databaseEngine)
-        ) {
-          return true;
-        }
-        return false;
-      });
-
-      const regionName = getRegionName(region.code);
-      selectedTermList.forEach((term) => {
-        const key = `${dbInstance.name}-${region.code}`;
-        const newRow: DataRow = {
-          id: -1,
-          key: term.code,
-          childCnt: 1,
-          cloudProvider: dbInstance.cloudProvider,
-          name: dbInstance.name,
-          processor: dbInstance.processor,
-          cpu: dbInstance.cpu,
-          memory: dbInstance.memory,
-          engineType: term.databaseEngine,
-          commitment: { usd: term.commitmentUSD },
-          hourly: { usd: term.hourlyUSD },
-          leaseLength: term.payload?.leaseContractLength ?? "N/A",
-          // we store the region code for each provider, and show the user the actual region information
-          // e.g. AWS's us-east-1 and GCP's us-east-4 are refer to the same region (N. Virginia)
-          region: regionName,
-        };
-
-        if (dataRowMap.has(key)) {
-          const existedDataRowList = dataRowMap.get(key) as DataRow[];
-          newRow.id = existedDataRowList[0].id;
-          existedDataRowList.push(newRow);
-        } else {
-          rowCnt++;
-          newRow.id = rowCnt;
-          dataRowMap.set(key, [newRow]);
-        }
-      });
-    });
-
-    dataRowMap.forEach((val) => {
-      val.sort((a, b) => {
-        // sort the row according to the following criteria
-        // 1. charge type
-        // 2. ascend by lease length
-        if (a.leaseLength == "N/A") {
-          return -1;
-        } else if (b.leaseLength == "N/A") {
-          return 1;
-        }
-        return Number(a.leaseLength[0]) - Number(b.leaseLength[0]);
-      });
-      val[0].childCnt = val.length;
-      dataRowList.push(...val);
-    });
-
-    // filter by keyword, we only enable this when the keyword is set by user
-    const keyword = config.keyword?.toLowerCase();
-    if (keyword) {
-      const filteredDataRowList: DataRow[] = dataRowList.filter((row) => {
-        if (
-          row.name.toLowerCase().includes(keyword) ||
-          row.memory.toLowerCase().includes(keyword) ||
-          row.processor.toLowerCase().includes(keyword) ||
-          row.region.toLowerCase().includes(keyword)
-        ) {
-          return true;
-        }
-        return false;
-      });
-      state.dataRow.push(...filteredDataRowList);
-      return;
-    }
-
-    state.dataRow.push(...dataRowList);
-  });
-};
 
 const clearAll = () => {
   searchConfigStore.clearAll();
-  state.dataRow = [];
+  dataTableStore.clearAll();
 };
 
 onMounted(() => {
   state.availableRegions = dbInstanceStore.getAvailableRegionList();
-  refreshDataTable();
+  dataTableStore.refresh();
 });
 </script>
